memory/requirements.txt
fastapi
uvicorn
python-dotenv  # For dotenv support
google-genai  # For google genai module
pydantic